trivial changes in several modules

+++ ppc/ppcd.h
@@ -6,10 +6,10 @@
 
 typedef signed   char       s8;
 typedef signed   short      s16;
-typedef signed   long       s32;
+typedef signed   int        s32;
 typedef unsigned char       u8;
 typedef unsigned short      u16;
-typedef unsigned long       u32;
+typedef unsigned int        u32;
 typedef float               f32;
 typedef double              f64;
